requirements.txt
requests>=2.31.0
dnspython>=2.6.1
httpx>=0.27.0
beautifulsoup4>=4.12.0
